refactor(dursto): move integration test helpers in new module
