+++ durable-storage/src/lib.rs
@@ -39,3 +39,4 @@ pub mod persistence_layer;
 pub mod registry;
 pub mod repo;
 pub mod storage;
+pub mod test_helpers;
